some workshop fixes
polygonal is more of an ongoing experiment as I learn more about stripper. I replaced the uppermost of 3 gravity triggers near the start, with a trigger that resets gravity. Later I want to change that back and add a trigger at the end of those zones to do that instead.