Mapper credits, a scope change
Large changelog

Right off the bat, the bad news is somewhere down the line my files were mildly mixed up so I DID lose a number of maps that were zoned/tiered. This will be a priority alongside the mapper credits.

Why release the update unfinished? Partially because I'm welcoming help while I work on this. I wanted to at least lay out an example of what I'm looking to do next with this project.

Zones WILL be revisited very soon, I've had a lot going on but now that things are settling down I can allot more time to cleaning up the repo.

At the moment the tier and zones sql's might be a little out of sync but that'll be resolved with the big re-zoning/testing coming.

------------------------------------
PR's and issue reports are more than welcome either here on git or in discord. Sayt ✿#0621