Folder structure update, MD5 hashes added
Refined the folder structure of the repo to be a little more organized. The two non-SQL files are now at the root. Everything else is folder managed.

Added is a csv file with ALL my maps I personally have in my collection of maps, their MD5 hash, and their file size in bytes. If I have zone issue reports I'll be asking for map MD5's to ensure the same version of the map is being used!

This is also potentially the start of a side project to help identify and further categorize the various versions of maps like utopia that are out there, help resolve map download conflicts on server(s).